Internal R6 object that process the hyper param grid generated in R

A R6 object that manage the export of metadata and parquet file of the hyper param grid generated in R

Details

This object works by:

inherits from the BaseEx class.

accept a hyper param grid generated with rpwf_grid_gen_.

calculate the hash of the grid and check the hash is in the database.

if the hash is found in the database:

assign the pin_name associated with the hash of the grid to self$pin_name.

check if the grid exists with self$pin_name.

if not exists, assign the grid to self$df.

if the file exists, then assign NULL to the self$df attribute to stop self$export_parquet() from executing.

if the hash is not found in the database:

assign the grid to self$df.

generate a new pin_name to write the grid.

generate a SQL query to update the database with the new hash and new pin_name

update the database with the generated SQL query with self$export_db()

write the parquet with self$export_parquet()

Super class

rpwf::BaseEx -> RGrid

Public fields

grid_obj

(character())
rpwf_grid_gen_() performs necessary clean ups before the grid can be used in python

Methods

Public methods

RGrid$new()

RGrid$set_attrs()

RGrid$clone()

Inherited methods

rpwf::BaseEx$exec_query()

rpwf::BaseEx$export()

rpwf::BaseEx$export_db()

rpwf::BaseEx$export_parquet()

rpwf::BaseEx$export_prep()

rpwf::BaseEx$find_pin_name_in_db()

rpwf::BaseEx$set_df()

rpwf::BaseEx$set_export_query()

rpwf::BaseEx$set_hash()

rpwf::BaseEx$set_pin_name()

Method new()

Create a new instance of the RGrid class. These are grids made by functions such as dials::grid_latin_hypercube and dials::grid_regular. See BaseEx for details of the attributes.

Usage

RGrid$new(grid_obj, db_con)

Arguments

grid_obj

(rpwf::rpwf_grid_gen_())
rpwf_grid_gen_() performs necessary clean ups before the grid can be used in python.

db_con

(DbCon)
a DbCon object, generated by rpwf_connect_db().

Method set_attrs()

Refresh the attributes using the current hash. Needed to be run because it updates the attributes with information from the Db

Usage

RGrid$set_attrs()

Method clone()

The objects of this class are cloneable with this method.

Usage

RGrid$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.